自定义键值映射

+++ src/main/java/com/example/databasedesign/annotation/ExcelDirectory.java
@@ -17,8 +17,7 @@ public @interface ExcelDirectory {
 
     public String name() default ""; //名称
 
-    public double width() default 25; //宽度默认是25 * 256
-
+    public int width() default 25; //宽度默认是25 * 256
 
     public boolean useInk() default false; //默认超链接
 }
